Move mailx from Recommends to Suggests
In Ubuntu, an MTA isn't necessarily installed by default; mailx (or at
least bsd-mailx, which provides mailx) requires one, forcing the user to
configure an MTA that is unlikely to be used. So Ubuntu moves mailx to
Suggests already.

This makes sense for Debian too. Users with an MTA installed will get
the additional functionality of the sysadmin being emailed from
debian-start.inc.sh; users without an MTA installed will not.